extracted ModerationPanel and Details to their own files
